Break si-frontend-mv-types-macros into smaller modules
